requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.5.2
click==8.1.8
exceptiongroup==1.3.1
fastapi==0.124.4
filelock==3.16.1
fire==0.7.1
fsspec==2025.3.0
h11==0.16.0
idna==3.11
imageio==2.35.1
Jinja2==3.1.6
joblib==1.4.2
lazy_loader==0.4
MarkupSafe==2.1.5
medmnist==3.0.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
packaging==26.0
pandas==2.0.3
pillow==10.4.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2026.1.post1
PyWavelets==1.4.1
PyYAML==6.0.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
sniffio==1.3.1
starlette==0.44.0
sympy==1.13.3
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2023.7.10
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.3
typing_extensions==4.13.2
tzdata==2026.1
uvicorn==0.33.0
